♻️ Refactor: eslint 적용 및 jsdoc 명세 추가
