+++ package.json
@@ -16,6 +16,11 @@
   "devDependencies": {
     "esbuild": "^0.24.2",
     "eslint": "^8.57.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "prettier": "^3.2.5"
   }
 }
